YapDatabase compiles on Windows
A combination of compiler settings and GNUstep method declarations mean we need to be more specific about certain casts and in particular `-[NSDictionary enumerateKeysAndObjectsUsingBlock:]` expects the block to be compatible with the generic type. Without providing type arguments, it expects the block to receive `(id<NSCopying>, id, BOOL *)`, producing a compiler error about incompatible block types

+++ YapDatabase/Utilities/YapBidirectionalCache.m
@@ -463,7 +463,7 @@ const YapBidirectionalCacheCallBacks kYapBidirectionalCacheDefaultCallBacks = (Y
 				leastRecentCacheItem = newItem;
 			
 			YDBLogVerbose(@"key(%@) <- new mostRecent [%ld of %lu]",
-			              key, CFDictionaryGetCount(key_obj_dict), (unsigned long)countLimit);
+			              key, (long)CFDictionaryGetCount(key_obj_dict), (unsigned long)countLimit);
 		}
 	}
